Inline base64 Hebrew font
Previously we were waiting to load fonts until the presence of a Hebrew
character triggered the custom.css font style, which only occured after
canopy.js was finished loading. Now, we put a base 64 encoded string in
the original index.html so that it must be loaded by the time canopy.js
initializes the page.